fix(performance): Display thresholds in all summary tabs
Display the threshold and metric in all the transaction
summary tabs. Perviously it only showed up in the main
summary page.
