+++ static/app/views/performance/transactionSummary/content.tsx
@@ -70,9 +70,6 @@ type Props = {
   onChangeFilter: (newFilter: SpanOperationBreakdownFilter) => void;
   onChangeThreshold?: (threshold: number, metric: TransactionThresholdMetric) => void;
   spanOperationBreakdownFilter: SpanOperationBreakdownFilter;
-  transactionThreshold?: number;
-  transactionThresholdMetric?: TransactionThresholdMetric;
-  loadingThreshold?: boolean;
 };
 
 type State = {
@@ -219,9 +216,6 @@ class SummaryContent extends React.Component<Props, State> {
       onChangeFilter,
       onChangeThreshold,
       spanOperationBreakdownFilter,
-      transactionThreshold,
-      transactionThresholdMetric,
-      loadingThreshold,
     } = this.props;
     const hasPerformanceEventsPage = organization.features.includes(
       'performance-events-page'
@@ -334,9 +328,6 @@ class SummaryContent extends React.Component<Props, State> {
           hasWebVitals={hasWebVitals}
           handleIncompatibleQuery={this.handleIncompatibleQuery}
           onChangeThreshold={onChangeThreshold}
-          transactionThreshold={transactionThreshold}
-          transactionThresholdMetric={transactionThresholdMetric}
-          loadingThreshold={loadingThreshold}
         />
         <Layout.Body>
           <StyledSdkUpdatesAlert />
