+++ static/app/views/performance/transactionSummary/header.tsx
@@ -2,7 +2,6 @@ import * as React from 'react';
 import styled from '@emotion/styled';
 import {Location} from 'history';
 
-import {openModal} from 'app/actionCreators/modal';
 import Feature from 'app/components/acl/feature';
 import {GuideAnchor} from 'app/components/assistant/guideAnchor';
 import Button from 'app/components/button';
@@ -25,10 +24,8 @@ import {tagsRouteWithQuery} from './transactionTags/utils';
 import {vitalsRouteWithQuery} from './transactionVitals/utils';
 import KeyTransactionButton from './keyTransactionButton';
 import TeamKeyTransactionButton from './teamKeyTransactionButton';
-import TransactionThresholdModal, {
-  modalCss,
-  TransactionThresholdMetric,
-} from './transactionThresholdModal';
+import TransactionThresholdButton from './transactionThresholdButton';
+import {TransactionThresholdMetric} from './transactionThresholdModal';
 import {transactionSummaryRouteWithQuery} from './utils';
 
 export enum Tab {
@@ -50,9 +47,6 @@ type Props = {
   handleIncompatibleQuery: React.ComponentProps<
     typeof CreateAlertFromViewButton
   >['onIncompatibleQuery'];
-  transactionThreshold?: number;
-  transactionThresholdMetric?: TransactionThresholdMetric;
-  loadingThreshold?: boolean;
 };
 
 class TransactionHeader extends React.Component<Props> {
@@ -140,33 +134,9 @@ class TransactionHeader extends React.Component<Props> {
     );
   }
 
-  openModal() {
-    const {
-      organization,
-      transactionName,
-      eventView,
-      transactionThreshold,
-      transactionThresholdMetric,
-      onChangeThreshold,
-    } = this.props;
-    openModal(
-      modalProps => (
-        <TransactionThresholdModal
-          {...modalProps}
-          organization={organization}
-          transactionName={transactionName}
-          eventView={eventView}
-          transactionThreshold={transactionThreshold}
-          transactionThresholdMetric={transactionThresholdMetric}
-          onApply={onChangeThreshold}
-        />
-      ),
-      {modalCss, backdrop: 'static'}
-    );
-  }
-
   renderSettingsButton() {
-    const {organization, loadingThreshold} = this.props;
+    const {organization, location, transactionName, eventView, onChangeThreshold} =
+      this.props;
 
     return (
       <Feature
@@ -179,12 +149,12 @@ class TransactionHeader extends React.Component<Props> {
               target="project_transaction_threshold_override"
               position="bottom"
             >
-              <Button
-                onClick={() => this.openModal()}
-                data-test-id="set-transaction-threshold"
-                icon={<IconSettings />}
-                disabled={loadingThreshold}
-                aria-label={t('Settings')}
+              <TransactionThresholdButton
+                location={location}
+                organization={organization}
+                transactionName={transactionName}
+                eventView={eventView}
+                onChangeThreshold={onChangeThreshold}
               />
             </GuideAnchor>
           ) : (
